requirements.txt
modal-client
pydantic
fastapi